Fix warning.

+++ libusb/os/darwin_usb.c
@@ -286,7 +286,7 @@ static void *event_thread_main (void *arg0) {
   if (kresult != kIOReturnSuccess) {
     usbi_err (ctx, "could not add hotplug event source: %s", darwin_error_str (kresult));
 
-    pthread_exit ((void *)kresult);
+    pthread_exit ((void *) (size_t) kresult);
   }
 
   /* arm notifiers */
